cmd-loadgen: initial implementation.
Lift over CRI-related bits from the cri-o NRI test code. Use it to
implement a test client that runs a number of parallel batches, each
creating, starting, stopping, and removing a set of pods with a set
of containers. The client measures, collects and can also save, the
latency of each CRI operation and the total test runtime. It reports
the min. and max. latencies, mean and standard deviation of latencies,
and {25,50,75,95}-percentiles of latencies for all operations.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>